Changed the appium install script

+++ install_appium_lib.sh
@@ -4,8 +4,12 @@ cd /tmp
 if [[ -e ruby_lib ]]; then
 	rm -r ruby_lib
 fi
-git clone https://github.com/EinarsNG/ruby_lib.git
-cd ruby_lib
-rake install
-cd ..
-rm -r ruby_lib
+if [[ -n $(gem list | grep rubocop) ]]; then
+	git clone https://github.com/EinarsNG/ruby_lib.git
+	cd ruby_lib
+	rake install
+	cd ..
+	rm -r ruby_lib
+else
+	echo Install Rubocop first: gem install rubocop
+fi
